fix: keep closeout proof on its executable test owners
