+++ scripts/verify-stable-main-closeout.mjs
@@ -2,8 +2,11 @@
 
 import { execFileSync } from "node:child_process";
 import { readFileSync, writeFileSync } from "node:fs";
-import { resolve } from "node:path";
-import { verifyStableMainCloseout } from "./lib/stable-release-closeout.mjs";
+import { basename, resolve } from "node:path";
+import {
+  verifyReleaseEvidenceChecksum,
+  verifyStableMainCloseout,
+} from "./lib/stable-release-closeout.mjs";
 
 function parseArgs(argv) {
   const values = new Map();
@@ -51,6 +54,19 @@ function gitSha(dir) {
 }
 
 function main() {
+  if (process.argv[2] === "verify-checksum") {
+    if (process.argv.length !== 4) {
+      throw new Error("usage: verify-stable-main-closeout.mjs verify-checksum <evidence-file>");
+    }
+    const path = resolve(process.argv[3]);
+    verifyReleaseEvidenceChecksum({
+      assetName: basename(path),
+      assetBytes: readFileSync(path),
+      checksum: readFileSync(`${path}.sha256`, "utf8"),
+    });
+    console.log(`release evidence checksum verified: ${basename(path)}`);
+    return;
+  }
   const args = parseArgs(process.argv.slice(2));
   const mainDir = resolve(args["main-dir"]);
   const tagDir = resolve(args["tag-dir"]);
